requirements.txt
pandas
numpy
scikit-learn
umap-learn
matplotlib
seaborn
python-dotenv
pyyaml
minio